Htmlentities added (not working fully)

+++ action/action_remove_dish.php
@@ -4,23 +4,27 @@
     require_once(__DIR__ . '/../database/connection.db.php');
     require_once(__DIR__ . '/../utils/session.php');
 
-    $id_dish = $_GET['id'];
+    if (isset($_GET['id'])) {
 
-    $session = new Session();
+        $id_dish = $_GET['id'];
 
-    $db = getDatabaseConnection();
+        $session = new Session();
 
-    $query = 'DELETE FROM DishType WHERE DishType.id_dish=?';
+        $db = getDatabaseConnection();
 
-    $stmt = $db->prepare($query);
+        $query = 'DELETE FROM DishType WHERE DishType.id_dish=?';
 
-    $stmt->execute(array($id_dish));
+        $stmt = $db->prepare($query);
 
-    $query = 'DELETE FROM Dish WHERE Dish.id=?';
+        $stmt->execute(array($id_dish));
 
-    $stmt = $db->prepare($query);
+        $query = 'DELETE FROM Dish WHERE Dish.id=?';
 
-    $stmt->execute(array($id_dish));
+        $stmt = $db->prepare($query);
+
+        $stmt->execute(array($id_dish));
+
+    }
 
     header('Location: '. $_SERVER['HTTP_REFERER']);
